requirements.txt
asgiref==3.6.0
Babel==2.12.1
certifi==2023.5.7
charset-normalizer==3.1.0
crispy-bootstrap5==0.7
Django==3.2.19
django-crispy-forms==2.0
django-environ==0.10.0
django-filter==23.2
django-phonenumber-field==7.1.0
django-user-agents==0.4.0
djangorestframework==3.14.0
hashids==1.3.1
idna==3.4
import-string==0.1.0
packaging==23.1
phonenumbers==8.13.11
python-dateutil==2.8.2
pytz==2023.3
requests==2.30.0
six==1.16.0
sqlparse==0.4.4
tenacity==8.2.2
tzdata==2023.3
ua-parser==0.16.1
urllib3==2.0.2
user-agents==2.2.0
